+++ clstd/clfifo.h
@@ -10,18 +10,18 @@ namespace clstd
   {
   private:
     u8*       m_buffer;   // the buffer holding the data
-    u32       m_size;     // the size of the allocated buffer
-    u32       m_in;       // data is added at offset (in % size)
-    u32       m_out;      // data is extracted from off. (out % size)
+    size_t    m_size;     // the size of the allocated buffer
+    size_t    m_in;       // data is added at offset (in % size)
+    size_t    m_out;      // data is extracted from off. (out % size)
     Locker*   m_pLocker;
   public:
     fifo();
     ~fifo();
   public:
-    b32   Initialize  (u32 size, b32 bLock);
-    u32   put         (const u8* data, u32 len);
-    u32   get         (u8* data, u32 len);
-    u32   size        ();
+    b32     Initialize  (size_t size, b32 bLock);
+    size_t  put         (const u8* data, size_t len);
+    size_t  get         (u8* data, size_t len);
+    size_t  size        ();
   };
 
 } // namespace clstd
