+++ clstd/smart/smartstream.h
@@ -152,7 +152,7 @@ public:
 
 
   };
-  typedef u32 (CALLBACK *IteratorProc)(iterator& it, clsize nRemain, u32_ptr lParam); // nRemain 是it.marker[0]到文件结尾的(剩余)长度. 目前不关心返回值
+  typedef size_t (CALLBACK *IteratorProc)(iterator& it, clsize nRemain, u32_ptr lParam); // nRemain 是it.marker[0]到文件结尾的(剩余)长度. 目前不关心返回值
   typedef const iterator const_iterator; 
 
 protected:
